Add files via upload
Using Statistical Models, analysis of the 'Lahman' package's Baseball related data sets, which are available on R, were conducted. With these statistical models, inferences can be made and predictive analysis can be performed to determine a range of Baseball Related phenomena, such as the expected mean salary of a team for a given year.